mqtt_client/requirements.txt
paho-mqtt
python-dotenv
requests>=2.0